requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
transformers
torch
scikit-learn
python-socketio
aiohttp
